requirements.txt
musicai_sdk
python-dotenv
yt-dlp
requests
ffmpeg

